% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosome_layouts.R
\name{add_chromosome_length}
\alias{add_chromosome_length}
\title{add chromosome lengths to a cn dataframe}
\usage{
add_chromosome_length(cn_df, version = c("hg19", "hg38"), chrom_col = "chr")
}
\arguments{
\item{cn_df}{dataframe of copy number information}

\item{version}{string. hg19 (default) or hg38}

\item{chrom_col}{string. name of column with chromosome information.}
}
\description{
Handles chromosomes specified with "chr#" or just #/X/Y.
}
