% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_filling.R
\name{add_missing_bins_for_cells}
\alias{add_missing_bins_for_cells}
\title{Add back chr, start, end of bins missing from cells}
\usage{
add_missing_bins_for_cells(
  state_df,
  version = c("hg19", "hg38"),
  bin_size = 5e+05,
  cell_metadata_cols = c()
)
}
\arguments{
\item{state_df}{the dataframe/tibble to insert the missing bins into}

\item{version}{which genome version you are using. Default "hg19", or select
"hg38"}

\item{bin_size}{size of bins that should be there.}

\item{cell_metadata_cols}{vector of columns to carry through to inferred bins}
}
\value{
input dataframe with extra bins for each cells that were missing.
}
\description{
Often we filter read bins based on various criteria, or tools we use might
drop bins of read/state data. This function will put missing bins back into
a dataframe for cells.
}
\details{
Basic operation will result in NAs for every column except cell_id, chr,
start, end for the newly added bins. But you can also carry over cell
metadata by specifying which columns to keep for the inferred bins. This
should be cell level data.
}
