% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosome_layouts.R
\name{add_telomere_positions}
\alias{add_telomere_positions}
\title{add telomere positions to a CN df}
\usage{
add_telomere_positions(cn_df, version = c("hg19", "hg38"))
}
\arguments{
\item{cn_df}{dataframe of CN information, segments or read bins. Just needs
a 'chr' column.}

\item{version}{string. hg19 (default) or hg38, not that it makes a
difference.}
}
\value{
input dataframe with new columns of telomere information
}
\description{
See \code{\link[=import_telos_file]{import_telos_file()}} but really just a convenience function.
Telomere positions in the UCSC gap file are all just 10Kb, except chr17,
which I set to 3Kb and 5Kb, again see linked function for details.
}
