% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_parsing.R
\name{add_tip_ancestors_to_df}
\alias{add_tip_ancestors_to_df}
\title{add parent node labels to a dataframe of cell ids with states}
\usage{
add_tip_ancestors_to_df(states_df, phylogeny)
}
\arguments{
\item{states_df}{bin based dataframe of cell ids and their states}

\item{phylogeny}{phylo object}
}
\value{
input dataframe, but with parent node information added. Creates a
'parent_node' column.
}
\description{
add parent node labels to a dataframe of cell ids with states
}
