% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_features.R
\name{add_wu_change_shape}
\alias{add_wu_change_shape}
\title{see dlptools::extract_wu_features}
\usage{
add_wu_change_shape(segs_df, change_split_val = 1)
}
\value{
tibble
}
\description{
see dlptools::extract_wu_features
}
