% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_segments.R
\name{align_segments}
\alias{align_segments}
\title{cut segments to equal widths across cells/samples.}
\usage{
align_segments(segs_df, chrom_col = "chr", cell_col = "cell_id")
}
\arguments{
\item{segs_df}{dataframe. Copy number segments of cells/samples. Need
chromosome, start, end.}

\item{chrom_col}{string. name of column containing chromosome information}

\item{cell_col}{string. name of column with cell/sample IDs}
}
\value{
tibble of broken down segments with same input columns + seg_width
of the new segments.
}
\description{
Segments are continuous stretches of a chromosome with the same copy number.
But, generally, these will be different across cells/samples. This function
will create segments in the smallest consistent intervals possible across
the samples. See example, but it's basically equalizing the segment widths
among all of the cells, so copy number is measured in the same "bin" across
cells.
}
\details{
The main value of this is that some tools (medicc, or plotting) want the same
intervals across all samples (i.e., bins of equal width). But going all the
way back to 500 kb bins (i.e., typical DLP) can be excessive when there are
longer segments within cells. Cutting segments into their minimal spans will
generally be a smaller set of data than a full 500kb bin breakdown (unless
CN data is highly noisy among samples).
}
\examples{
sgs <- tibble::tibble(
  cell_id = c("A", "A", "B", "B"),
  chr = rep("chr1", 4),
  start = c(1, 11, 1, 8),
  end = c(10, 25, 7, 25),
  state = c(2, 4, 3, 8)
)

sgs

align_segments(sgs)

}
