% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_plots.R
\docType{data}
\name{ASCN_COLORS}
\alias{ASCN_COLORS}
\title{colors for signals results of allele balances}
\format{
An object of class \code{character} of length 17.
}
\usage{
ASCN_COLORS
}
\description{
colors for signals results of allele balances
}
\keyword{datasets}
