% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{ASCN_PHASE_COLORS}
\alias{ASCN_PHASE_COLORS}
\title{ASCN phase colors}
\format{
An object of class \code{character} of length 5.
}
\usage{
ASCN_PHASE_COLORS
}
\description{
ASCN phase colors
}
\keyword{datasets}
