% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{basic_tile_plot}
\alias{basic_tile_plot}
\title{create a tile plot of read state calls.}
\usage{
basic_tile_plot(reads_df)
}
\arguments{
\item{reads_df}{a table of reads data (e.g., could load with
\code{\link[=import_dlp_files]{import_dlp_files()}})}
}
\value{
ggplot object
}
\description{
builds a basic ggplot with geom_tile on a reads df.
Only expects columns of cell_id, start, state, chr
}
