% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centromere_marking.R
\name{bin_within_centromere}
\alias{bin_within_centromere}
\title{mark edge of bin as within centromere or not}
\usage{
bin_within_centromere(
  reads_df,
  padding = 0,
  bin_location_column = "start",
  version = c("hg19", "hg38")
)
}
\arguments{
\item{reads_df}{tibble of read data}

\item{padding}{int of number of BP to add to each side of the centromere}

\item{bin_locatation_column}{which column to use as the location of a bin.
popular choices are the start of a bin, middle of a bin, or and of a bin.}

\item{hg19}{bool. Not needed unless this function is also adding the
centromere information to the data frame. See
dlptools::add_centromere_locations()
which will be called if the correct centromere columns are not present.}

\item{hg38}{bool. Same as hg19...but for hg38.}
}
\value{
input table, but with a boolean 'within_centro' column added (and
potentially other centromere information columns, if needed)
}
\description{
Can optionally specify a padding to mark locations of bins as being close
enough to a centromere. Often bins near centromeres are corrupt in their
state calls. In the past, we have filtered within 3 Mb of centromeres.
}
\details{
IMPORTANT! This adds the padding to each side of the centromere. So if you
specify a padding of 3 Mb, it will be within 3 Mb of the start and 3 Mb of
the end of the centromere.
}
