% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{bme_vec}
\alias{bme_vec}
\title{picks beginning, middle, and end of a vector
to handle when vectors that are too long are
passed}
\usage{
bme_vec(vec, vec_name)
}
\description{
picks beginning, middle, and end of a vector
to handle when vectors that are too long are
passed
}
