% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoint_sharing.R
\name{breakpoint_sharing}
\alias{breakpoint_sharing}
\title{summarize pairwise breakpoint sharing across cells}
\usage{
breakpoint_sharing(seg_df, jaccard = FALSE)
}
\arguments{
\item{seg_df}{dataframe. Segmented copy number profiles per cell.}

\item{jaccard}{bool. True will calculate and return jaccard similarity.}
}
\value{
pairwise matrix of the number of shared breakpoints between each
pair of cells, or the jaccard similarity.
}
\description{
Will summarize the number of shared breakpoints (changes in copy number) for
all pairs of cells. Alternatively, will return jaccard similarity between
all pairs (intersection of shared breakpoints over the union of breakpoints).

Returns a pairwise matrix that can be used for various downstream things,
like clustering.
}
