% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plot.R
\name{build_left_annot}
\alias{build_left_annot}
\title{builds the left-side annotations of the cells}
\usage{
build_left_annot(
  anno_df = NULL,
  anno_cols_list = list(),
  annotation_legend_param = list(),
  annotation_name_gp = 16,
  show_annotation_legend = TRUE,
  clones_df = NULL,
  only_largest_clone_group = FALSE,
  labels_fontsize = 8,
  clone_palette = NULL
)
}
\arguments{
\item{anno_df}{annotations dataframe with cell_id column and annotation for
each cell id to be added to a heatmap.}

\item{anno_cols_list}{list of named vectors specifying colors for annotations
example: list(passage=c(\code{1}='#2872bc', \code{19}='#d23e3e'))}

\item{annotation_name_gp}{Int: Fontsize for label names}

\item{clones_df}{dataframe of clone ideas (clone_id) for each cell_id. Both
columns required.}

\item{labels_fontsize}{how large to make text labels}

\item{clone_palette}{named vector of colors to give to clones. E.g., c(
A='#12345', B='#67890'). See dlptools::make_clone_palette()}
}
\description{
builds the left-side annotations of the cells
}
