% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoint_sharing.R
\name{calc_jaccard_similarity}
\alias{calc_jaccard_similarity}
\title{calculate jaccard similarity using breakpoint matrices}
\usage{
calc_jaccard_similarity(bps_mtx, shared_bps)
}
\arguments{
\item{bps_mtx}{sparse matrix. Breakpoint presence in cells}

\item{shared_bps}{sparse matrix. the numbers of shared breakpoints between
cells.}
}
\value{
sparse matrix of jaccard similarity
}
\description{
Jaccard is the number of shared breakpoints (intersection), over the total
number of unique breakpoints (union) between two cells.
}
