% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converting_segments_and_reads.R
\name{cap_dlp_to_chrom_lengths}
\alias{cap_dlp_to_chrom_lengths}
\title{cap DLP data at the lengths of chromosomes}
\usage{
cap_dlp_to_chrom_lengths(
  rs_df,
  chrom_col = "chr",
  genome_version = c("hg19", "hg38")
)
}
\arguments{
\item{rs_df}{tibble/dataframe. Reads or segments dataframe}

\item{chrom_col}{string. Name of chromosome column}

\item{genome_version}{string. Name of genome version to use. Default: hg19}
}
\value{
input dataframe with final bins/segments not exceeding total
chromosome lengths
}
\description{
DLP can output bins/segments longer than the actual chromosomes as it chops
to 500kb bins. This function sets the \code{end} column to be no longer than
the chromosome it occurs on.
}
