% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_plots.R
\name{cell_cn_profile}
\alias{cell_cn_profile}
\title{plot copy number profile plot of a cell.}
\usage{
cell_cn_profile(reads_df, cell_ids = c(), pseduo_log_y = FALSE)
}
\arguments{
\item{reads_df}{dataframe of read bins like information.}

\item{cell_ids}{string/vector of cell ids to plot. Remember, too many and
the plot will be useless. If blank, will plot all cells in dataframe, but
only if less than 10, because honestly at that point, you probably need a
different plot.}

\item{pseudo_log_y}{boolean. Pseudo-log transform the y axis, as sometimes
high copy numbers obscure the plot. Or you can always set a
ggplot2::coord_cartesian() yourself on the plot returned by this function.}
}
\value{
ggplot2 object
}
\description{
Basically a geom_point of the float copy number, colored by the integer
state number. singals::plotCNprofile() is similar, possibly with more
features. But this function works in a pinch.
}
