% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_hm.R
\name{cell_id_order_as_plotted}
\alias{cell_id_order_as_plotted}
\title{grab cell ids in the order that they are plotted}
\usage{
cell_id_order_as_plotted(phylo_tree)
}
\arguments{
\item{phylo_tree}{a phylo object of the tree being used in the heatmap}
}
\value{
vector of cell_ids (or really, whatever are the tip labels)
}
\description{
This is to align state calls and other things. It will make a ggplot of the
tree the same as the heatmap code does, and then pull out the cell ID in the
order that they are plotted
}
