% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_believability.R
\name{cell_states_to_strings}
\alias{cell_states_to_strings}
\title{collapse cell states to strings}
\usage{
cell_states_to_strings(
  states_df,
  states_col = "state",
  cell_id_col = "cell_id"
)
}
\arguments{
\item{states_df}{a dataframe with states for each bin for cells (or whatever
the tree tips are)}

\item{states_col}{name of the column with state values}

\item{cell_id_col}{name of the column with the cell ids (tree tip names)}
}
\description{
Converts a long format states dataframe into 1 row per cell with all
states converted to a letter string.
}
