% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreground_utils.R
\name{characterize_foreground_allele_state_changes}
\alias{characterize_foreground_allele_state_changes}
\title{mark types of allele specific foreground state changes}
\usage{
characterize_foreground_allele_state_changes(bin_states_df)
}
\description{
mark types of allele specific foreground state changes
}
