% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreground_utils.R
\name{characterize_foreground_total_state_changes}
\alias{characterize_foreground_total_state_changes}
\title{mark the types of change between a parent node state and the tip.}
\usage{
characterize_foreground_total_state_changes(
  bin_states_df,
  state_col = "state",
  parent_state_col = "parent_state"
)
}
\arguments{
\item{bin_states_df}{bin based dataframe with parent and tip state columns.}

\item{state_col}{name of tip state column}

\item{parent_state_col}{name of parent state column.}
}
\value{
input dataframe with new columns
}
\description{
This compares the tip and parent state columns to evaluate if a foreground
change has occurred, what type of change, and how much of a change.
}
\details{
e.g., if parent state is 5, tip is 4, it will get marked as a 'foreground'
event, fg-gain, with a change of 1.

4 coulmns are added:
foreground: boolean, if parent and tip states are different
fg_type: character, fg-gain, fg-loss, or NA
fg_change: numeric, tip - parent state. Positive indicates gain in tip.
background: boolean, inverse of foreground
}
