% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plot.R
\name{check_args}
\alias{check_args}
\title{confirms arguments are compatible for the plotting wrapper}
\usage{
check_args()
}
\arguments{
\item{all}{of the plot_state_hm args}
}
\description{
confirms arguments are compatible for the plotting wrapper
}
