% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreground_utils.R
\name{check_if_state_columns_are_valid}
\alias{check_if_state_columns_are_valid}
\title{internal check for using the characterization function}
\usage{
check_if_state_columns_are_valid(
  states_df,
  state_col = "state",
  parent_state_col = "parent_state"
)
}
\description{
To characterize forground changes, need wither numeric columns or factors
with a defined order to them.
}
