% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plot.R
\name{check_or_fetch_clone_palette}
\alias{check_or_fetch_clone_palette}
\title{confirm pallete given has enough colors or generate one}
\usage{
check_or_fetch_clone_palette(clones_df, clone_palette = NULL)
}
\arguments{
\item{clones_df}{dataframe with a column of clone_id and the cells
assigned to each (long format)}

\item{clone_palette}{named vector of how to color the clones.}
}
\description{
confirm pallete given has enough colors or generate one
}
