% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_believability.R
\name{check_the_vibe}
\alias{check_the_vibe}
\title{just a silly alias.}
\usage{
check_the_vibe(states_df, tree, states_col = "state", cell_id_col = "cell_id")
}
\description{
see dlptools::compute_tip_sibling_distances()
}
