% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chr_name_check}
\alias{chr_name_check}
\title{internal for checking the names of chromosome columns in a dataframe.}
\usage{
chr_name_check(df, exp_chr_name)
}
\arguments{
\item{df}{a dataframe being manipulated in some function}

\item{exp_chr_name}{a string of the chromosome column name the function is
expecting to see.}
}
\value{
bool TRUE if all good, FALSE with a message otherwise.
}
\description{
used for error catching.
}
