% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{data}
\name{CNV_COLOURS}
\alias{CNV_COLOURS}
\title{standard colors used in dlp plots}
\format{
An object of class \code{character} of length 13.
}
\usage{
CNV_COLOURS
}
\description{
standard colors used in dlp plots
}
\keyword{datasets}
