% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_differences.R
\name{compare_two_cells}
\alias{compare_two_cells}
\title{internal workhorse of pairwise_bin_difference function}
\usage{
compare_two_cells(cell_1_df, cell_2_df, min_seg_length)
}
\arguments{
\item{cell_1_df}{dataframe of bin state data for one cell}

\item{cell_2_df}{dataframe of bin state data for the other cell}

\item{min_seg_length}{int/double. minimum length of matching/non-matching
runs of bins to consider in the calculation of the differences between the
cells.}
}
\value{
1 row tibble summarizing differences.
}
\description{
internal workhorse of pairwise_bin_difference function
}
