% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{confirm_cols_present}
\alias{confirm_cols_present}
\title{confirm if columns exist in dataframe}
\usage{
confirm_cols_present(cols, df)
}
\arguments{
\item{cols}{vector of strings. Names of columns you are searching for in the
dataframe.}

\item{df.}{The dataframe to look in for the columns}
}
\value{
TRUE or raises a stop
}
\description{
confirm if columns exist in dataframe
}
