% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_heatmap.R
\name{convert_bins_to_columns_cells_to_rows}
\alias{convert_bins_to_columns_cells_to_rows}
\title{flip bins from rows to columns and put cell ids as rows}
\usage{
convert_bins_to_columns_cells_to_rows(states_df)
}
\arguments{
\item{states_df}{tibble/dataframe with bins and rows and cell ids as columns}
}
\value{
tibble of the inverse frame.
}
\description{
flip bins from rows to columns and put cell ids as rows
}
