% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_differences.R
\name{convert_dists_to_pairwise}
\alias{convert_dists_to_pairwise}
\title{convert cell distances to a pairwise matrix}
\usage{
convert_dists_to_pairwise(cell_dists)
}
\arguments{
\item{cell_dists}{dataframe from dlptools::pairwise_bin_difference()}
}
\value{
matrix of index vs comp cell distances.
}
\description{
Takes the output of dlptools::pairwise_bin_difference() and returns a
pairwise matrix of the distances. Useful to then pass to functions like
stats::hclust() and related ideas.
}
