% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_long_reads_to_wide}
\alias{convert_long_reads_to_wide}
\title{convert long format reads to wide format}
\usage{
convert_long_reads_to_wide(reads_df, state_col = "state")
}
\arguments{
\item{reads_df}{is the reads table to convert.}
}
\value{
wide format table
}
\description{
A common manipulation with reads files for various analyses is to reshape
long format reads data (each row is a 500kb bin with state values for each
cell) to wide format, with chr_start_end rows and cell_id columns.
}
\details{
minimal required columns for input are: chr,start,end,cell_id,state
}
