% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_segment.R
\name{convert_to_mtx}
\alias{convert_to_mtx}
\title{convert a dataframe to a wide matrix of bin states}
\usage{
convert_to_mtx(chr_df)
}
\value{
list of Y, the matrix, and cell_ord, the cell IDs in the matrix
}
\description{
convert a dataframe to a wide matrix of bin states
}
