% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_features.R
\name{count_default_process_feats}
\alias{count_default_process_feats}
\title{wrapper around classic process based features}
\usage{
count_default_process_feats(segs_df, sample_col = "cell_id", with_cn = FALSE)
}
\arguments{
\item{segs_df}{dataframe. Segmented CN for samples.}

\item{sample_col}{string. Column name of sample ID column.}

\item{with_cn}{bool. default FALSE. FALSE means do not include CN in the
feature counts, TRUE will include it.}
}
\description{
Extracts counts using the default set categories of the process based
features, including:
}
\details{
\enumerate{
\item \link{extract_segment_sizes}
\item \link{extract_changepoint}
\item \link{extract_bp_per_arm}
\item \link{extract_bp_per_window}
\item \link{extract_oscillations}
}

Not included is CN, which can be added by specifying with_cn=TRUE

Internally using furrr::future_map, so if you set up a future::plan(), this will execute in parallel. Recommended. See Example.
}
\examples{
\dontrun{
future::plan(future::multicore, workers = 10)
dlptools::count_default_process_feats(segs_df, sample_id = "cell_id")
}
}
