% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosome_layouts.R
\name{create_chrom_arm_intervals}
\alias{create_chrom_arm_intervals}
\title{create a list of intervals spanning chromosome arms}
\usage{
create_chrom_arm_intervals(genome_version = c("hg19", "hg38"))
}
\arguments{
\item{genome_version}{string. "hg19" (default) or "hg38"}
}
\value{
list. Named by chromosome, vectors of how to break a chromsome into
intervals of arms.
}
\description{
Splits a chromosome at the middle of the centromere. Sets up intervals for
splitting each chromosome arm.
}
