% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosome_layouts.R
\name{create_chrom_window_intervals}
\alias{create_chrom_window_intervals}
\title{break a chromosome up into intervals of a defined window size}
\usage{
create_chrom_window_intervals(window_size, genome_version = c("hg19", "hg38"))
}
\arguments{
\item{window_size}{integer. The size of window to split the chromosome into.}

\item{genome_version}{string. "hg19" (default) or "hg38"}
}
\value{
list. Named by chromosome, vectors of window starts.
}
\description{
break a chromosome up into intervals of a defined window size
}
