% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plot.R
\name{create_chromosome_column_fct}
\alias{create_chromosome_column_fct}
\title{create a sorted factor vector of chromosomes}
\usage{
create_chromosome_column_fct(states_mat)
}
\arguments{
\item{states_mat}{matrix of cell_id named rows and bin (chr_start_end)
columns}
}
\value{
factor of chromosomes with sorted levels
}
\description{
This is mostly just for splitting the the states heatmap by chromosome.
Will be naturally sorted.
}
