% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_filling.R
\name{create_expected_bins}
\alias{create_expected_bins}
\title{create bins of given size for a genome}
\usage{
create_expected_bins(version = c("hg19", "hg38"), bin_size = 5e+05)
}
\arguments{
\item{version}{default "hg19", or can select "hg38"}

\item{bins_size}{int of how big to make the bins}
}
\value{
tibble of bins for each chromosome. Columns: chr start end
}
\description{
create bins of given size for a genome
}
