% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_heatmap.R
\name{create_tree_copynumber_heatmap}
\alias{create_tree_copynumber_heatmap}
\title{create the classic tree-state heatmap plot}
\usage{
create_tree_copynumber_heatmap(
  phylo_tree,
  states_df,
  clones_df,
  annos_df = NULL,
  file_name = NULL,
  ...
)
}
\arguments{
\item{phylo_tree}{a phylo object of the tree to plot.}

\item{annos_df}{optional table with a \code{sample_id} column and any number of
annotations to add for those sample IDs.}

\item{file_name}{optional string of where to save a png image of the heatmap.}

\item{states}{a table of cells states with a column of cell_ids, and other
columns of state bins (chrom_start_end).}

\item{clones}{a table with at least cell_id and clone_id labels}
}
\value{
either a ComplexHeatmap::Heatmap().draw() object or noting and writes
a file.
}
\description{
This will create a ComplexHeatmap::Heatmap with a tree on the left, and a
heatmap of states, with clone ID labels. It can also add annotations to the
left side, if provided. (not ideal set up) but you can also pass arguments
for \code{\link[=make_copynumber_heatmap]{make_copynumber_heatmap()}} and  \code{\link[=plot_combined_heatmap]{plot_combined_heatmap()}} if needed. See
those functions for arguments not specified here.
}
\details{
This function can either return a heatmap or dump it to a file, if a
file_name is provided.
}
