% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cust_mode}
\alias{cust_mode}
\title{find most common element in a vector}
\usage{
cust_mode(x, na_rm = FALSE)
}
\arguments{
\item{x}{vector of values}
}
\value{
most common element of input vector
}
\description{
For ties, just the first one in the list is returned.
}
