% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cut_categories_and_count}
\alias{cut_categories_and_count}
\title{count values that map to given categories}
\usage{
cut_categories_and_count(df, targ_col, sample_col, breaks, labels)
}
\arguments{
\item{df}{dataframe. Contains a column of values you want to count per sample}

\item{targ_col}{string. Column you want to put in categories and counted.}

\item{sample_col}{string. Name of the column with sample ids.}

\item{breaks}{vector of doubles. Bounds for the categories.}

\item{labels}{vector of strings. What to call the categories.}
}
\value{
tibble
}
\description{
count values that map to given categories
}
