% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oscilations.R
\name{decide_if_oscillation}
\alias{decide_if_oscillation}
\title{decide if a 3 segment chain counts as an oscillation}
\usage{
decide_if_oscillation(
  left_cn,
  middle_cn,
  right_cn,
  middle_bound = Inf,
  ends_bound = 0
)
}
\arguments{
\item{left_cn}{int. CN value of segment on right side of the chain}

\item{middle_cn}{int. CN value of segment middle of the chain}

\item{middle_bound}{int|Inf. Default Inf. How different the middle CN is
allowed to be from either side to count as an oscillation. See
\link{extract_oscillations} for a discussion of this}

\item{ends_bound}{int. Default 0. How different the ends of a 3-segment set
are allowed to be to count as an oscillation}
}
\value{
boolean. True if 3 CNs constitute an oscillation, False if not.
}
\description{
See \link{extract_oscillations} for a discussion of oscillations and
recognizing them in sample copy number segments.
}
