% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{DEFAULT_CONTINUOUS_COLOR_RANGE}
\alias{DEFAULT_CONTINUOUS_COLOR_RANGE}
\title{min, median, max for a continuous color range}
\usage{
DEFAULT_CONTINUOUS_COLOR_RANGE
}
\description{
min, median, max for a continuous color range
}
