% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{does_palette_cover_plot_vals}
\alias{does_palette_cover_plot_vals}
\title{internal to confirm color choice was ok}
\usage{
does_palette_cover_plot_vals(p_vals, palette, exclude_na = TRUE)
}
\arguments{
\item{p_vals}{vector of values to be plotted in the heatmap with the palette}

\item{palette}{palette being used. A named vector with names corresponding
to the plot values.}

\item{exclude_na}{bool. Whether to drop NAs from the check. ComplexHeatmap
handles NAs with a default color regardless of palette.}
}
\description{
internal to confirm color choice was ok
}
