% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_filling.R
\name{expand_length_to_bins}
\alias{expand_length_to_bins}
\title{For a given length, create bins of a given size}
\usage{
expand_length_to_bins(total_len, bin_size = 5e+05)
}
\arguments{
\item{total_len}{int to create bins for}

\item{bin_size}{int of size of bin}
}
\value{
tibble of bins with columns of start and end
}
\description{
e.g., length = 10, bin = 5, will get bins: 1-5, 6-10
}
