% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_features.R
\name{extract_bp_per_arm}
\alias{extract_bp_per_arm}
\title{convenience function to extracting breakpoints per chromosome arm}
\usage{
extract_bp_per_arm(
  segs_df,
  sample_col = "cell_id",
  chrom_col = "chr",
  genome_version = c("hg19", "hg38"),
  return = c("values", "counts")
)
}
\value{
dataframe. Sample IDs and the observed breakpoint counts on
chromosome arms.
}
\description{
just calling the generic \link{extract_breakpoints}, with some
pre-loaded options. See that function for details.
}
