% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_features.R
\name{extract_bp_per_window}
\alias{extract_bp_per_window}
\title{convenience function to extracting breakpoints per window size}
\usage{
extract_bp_per_window(
  segs_df,
  window_size = 1e+07,
  sample_col = "cell_id",
  chrom_col = "chr",
  genome_version = c("hg19", "hg38"),
  return = c("values", "counts")
)
}
\value{
dataframe. Sample IDs and the observed breakpoint counts observed in
the specified window size.
}
\description{
just calling the generic \link{extract_breakpoints}, with some
pre-loaded options. See that function for details.
}
