% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_features.R
\name{extract_breakpoints}
\alias{extract_breakpoints}
\title{extracting counts of breakpoints per user-defined scope}
\usage{
extract_breakpoints(
  segs_df,
  scope = c("windows", "arms"),
  genome_version = c("hg19", "hg38"),
  window_size = 1e+07,
  sample_col = "cell_id",
  chrom_col = "chr",
  return = c("values", "counts")
)
}
\arguments{
\item{segs_df}{dataframe. Copy number segments for samples}

\item{scope}{string. "windows" (default) or "arms", i.e., what to target for
counting}

\item{genome_version}{string. "hg19" (default) or "hg38"}

\item{window_size}{integer. How big of a window to use, if extracting counts
per a "windows" scope. Most publications use 10Mb, which is the default
(10e6)}

\item{sample_col}{string. Name of column with cell/sample names}

\item{chrom_col}{string. Name of column with chromosome names}

\item{return}{string. "values" (default) or "counts". Values are the
observed values for cells, counts are the counts of these values in
pre-determined categories.}
}
\value{
dataframe. Sample IDs and the observed breakpoint counts per scope.
}
\description{
Counting the number of breakpoints (i.e., transitions between copy number
segments) per arm or per window (typically 10Mb).
}
