% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_features.R
\name{extract_cn}
\alias{extract_cn}
\title{extract copy state values.}
\usage{
extract_cn(
  segs_df,
  sample_col = "cell_id",
  state_col = "state",
  return = c("values", "counts")
)
}
\arguments{
\item{segs_df}{dataframe. copy number segments for samples.}

\item{sample_col}{string. Name of column with CN state values}

\item{return}{string. "values" (default) or "counts". Values are the
observed values for cells, counts are the counts of these values in
pre-determined categories.}
}
\description{
This function is really only here to offer the counting in pre-set
categories, and provide alignment with other feature extraction types. It is
mostly a pointless function and you should probably do this yourself.
}
\details{
Also, it is a question if copy state should even be used in
signature fitting, as this can occassionally lead to the same process being
artifically split by ploidy of samples. See
\href{https://www.nature.com/articles/s41586-022-04789-9}{Drews et al.}
for a discussion of this idea.
}
