% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{factor_column_mixedsort}
\alias{factor_column_mixedsort}
\title{naturally sort a column from a dataframe.}
\usage{
factor_column_mixedsort(df, col)
}
\arguments{
\item{df}{a dataframe that contains the column to sort}

\item{col}{the column to sort}
}
\value{
factor of the column with sorted levels
}
\description{
Common plotting issue, used mixed sort so things like, e.g., chromosomes
get sorted properly chr1, chr2, ..., chr10
}
