% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{fetch_continuous_color_ramp}
\alias{fetch_continuous_color_ramp}
\title{internal function for setting up heatmap continuous range colors
chooses defaults, unless overwritten by user.}
\usage{
fetch_continuous_color_ramp(plotting_values, custom_continuous_range = NULL)
}
\description{
internal function for setting up heatmap continuous range colors
chooses defaults, unless overwritten by user.
}
