% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_differences.R
\name{find_nearest_neighbours}
\alias{find_nearest_neighbours}
\title{find the nearest neighbour of each cell}
\usage{
find_nearest_neighbours(pairwise_diffs)
}
\arguments{
\item{pairwise_diffs}{tibble output of \code{\link[=pairwise_bin_difference]{pairwise_bin_difference()}}}
}
\value{
tibble/dataframe
}
\description{
After running \code{\link[=pairwise_bin_difference]{pairwise_bin_difference()}}, we can feed that output
dataframe to this function to find the nearest neighbour of each cell. The
pairwise function does the comparisons in a non-redundant way (A-B, A-C,
B-C), leaving out some redundant comparisons (i.e., C-A, C-B). If A's
nearest neighbour is C, it's not guaranteed that C's nearest neighbour is A.
This function takes a cell specific focus and finds the nearest neighbour of
each.
}
\details{
As with \code{\link[=pairwise_bin_difference]{pairwise_bin_difference()}}, using \code{\link[furrr:future_map]{furrr::future_map()}}
internally, so speed improvements with a \code{\link[future:plan]{future::plan()}} being set.
}
