% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_heatmap.R
\name{format_sitka_tree}
\alias{format_sitka_tree}
\title{clean tree tip labels and drop any locus tips from sitka trees}
\usage{
format_sitka_tree(tree)
}
\arguments{
\item{tree}{phylo object as read by ape::read.tree}
}
\value{
phylo object cleaned of "cell_" notation
}
\description{
'Locus tips' are from sitka and are locus values that end up on the
tip of trees. Also removes the "cell_" prefix from tip labels, which is
also a consequence of sitka.
}
