% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plot.R
\name{format_states_for_hm}
\alias{format_states_for_hm}
\title{format states for plotting in a heatmap}
\usage{
format_states_for_hm(
  states_df,
  state_col,
  continuous = FALSE,
  replace_11_with_11plus = TRUE
)
}
\arguments{
\item{states_df}{long format states information}

\item{state_col}{the column being plotted in the heatmap}
}
\value{
tibble (or maybe just a matrix ready to go?)
}
\description{
format states for plotting in a heatmap
}
