% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_heatmap.R
\name{get_clone_members}
\alias{get_clone_members}
\title{create list of list of cell_ids that belong to each clone id.}
\usage{
get_clone_members(clones_df)
}
\arguments{
\item{clones_df}{a tibble/df with cell_id, clone_id columns}
}
\value{
named list of lists, with clone_id as names
}
\description{
create list of list of cell_ids that belong to each clone id.
}
