% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_column_metrics}
\alias{get_column_metrics}
\title{get plotted values bounds}
\usage{
get_column_metrics(vals, min_max = FALSE)
}
\description{
min, max, median of a column to generate a color palette for
}
