% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_package_file_path}
\alias{get_package_file_path}
\title{retrieve file path to package data files}
\usage{
get_package_file_path(file_name)
}
\arguments{
\item{file_name}{string of the file to target for loading}
}
\value{
file path string
}
\description{
retrieve file path to package data files
}
