% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_parsing.R
\name{get_tip_parents}
\alias{get_tip_parents}
\title{resolve who the immediate parent nodes are for each tip.}
\usage{
get_tip_parents(phylogeny)
}
\arguments{
\item{phylogeny}{phylo object of tree of interest.}
}
\value{
tibble of each tip and it's immediate parent node label.
}
\description{
resolve who the immediate parent nodes are for each tip.
}
