% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_believability.R
\name{get_tips_that_avoid_redundant_comps}
\alias{get_tips_that_avoid_redundant_comps}
\title{get tips labels that will avoid duplicate sibling comparisons}
\usage{
get_tips_that_avoid_redundant_comps(tree)
}
\arguments{
\item{tree}{a phylo object}
}
\value{
vector of tips labels that will lead to non-redundant
}
\description{
I.e., in a tree: (A, (B, C)) we don't want to compare B to C
and then C to B, we only need to do one of those comparisons.
}
