% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_heatmap.R
\name{import_annotations_df}
\alias{import_annotations_df}
\title{read an annotations file (t(c)sv)}
\usage{
import_annotations_df(annotations_file)
}
\arguments{
\item{annotations_file}{path to file. Requires at least sample_id column}
}
\value{
tibble
}
\description{
read an annotations file (t(c)sv)
}
