% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plot.R
\name{import_clones}
\alias{import_clones}
\title{read a clones t(c)sv file}
\usage{
import_clones(clones_f)
}
\arguments{
\item{clones_f}{a path to a t(c)sv file with cell_id and clone_id columns}
}
\value{
tibble
}
\description{
read a clones t(c)sv file
}
