% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing.R
\name{import_dlp_files}
\alias{import_dlp_files}
\title{generic function to import dlp files}
\usage{
import_dlp_files(dlp_sc_dir, file_type = c("metrics", "segs", "reads"))
}
\arguments{
\item{dlp_sc_dir}{the parent directory where the dlp outputs are saved (see
the README.md for expected structure)}

\item{file_type}{options are: "metrics", "segs", "reads"}
}
\value{
a tibble of the information from all dlp outputs combined
}
\description{
Will create a dataframe of all files of a certain type. This is for the three
common dlp files that you might want to load: metrics, segments, or reads.
}
