% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plot.R
\name{import_tree}
\alias{import_tree}
\title{read a tree from a file in newick format}
\usage{
import_tree(tree_f)
}
\arguments{
\item{tree_f}{a path to a tree file}
}
\value{
phylo object
}
\description{
read a tree from a file in newick format
}
