% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_heatmap.R
\name{import_wideformat_states_file}
\alias{import_wideformat_states_file}
\title{import wide format states data and flip to cell rows and bin columns.}
\usage{
import_wideformat_states_file(states_f)
}
\arguments{
\item{states_f}{path to a state calls file}
}
\value{
tibble with cell_id rows and bin columns
}
\description{
This is a file with the first column are bins (chrom_start_end) and the
other columns are cell ids. Internally, the function will flip this.
}
