% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreground_utils.R
\name{is_not_numeric_lacks_order}
\alias{is_not_numeric_lacks_order}
\title{confirms a vector is not numeric and is unorderd}
\usage{
is_not_numeric_lacks_order(vec)
}
\description{
returns TRUE if the vector is not numeric and is missing a defined order.
Used as a check of suitability for comparing events between tip and parent
nodes.
}
