% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_segment.R
\name{joint_seg_chromosome}
\alias{joint_seg_chromosome}
\title{Perform joint segmentation on a single chromosome}
\usage{
joint_seg_chromosome(
  chr_df,
  chrom_col = "chr",
  max_bps = NULL,
  bin_fraction = 0.5
)
}
\arguments{
\item{chr_df}{dataframe. State calls of the DLP bins data.}

\item{chrom_col}{string. Name of the chromosome column}

\item{max_bps}{int. Can specify the maximum number of breakpoints to find}

\item{bin_fraction}{float. What fraction of the bins to target as the maximum number of segments.}
}
\value{
dataframe. Cell segments that match the jointly found segments.
}
\description{
Perform joint segmentation on a single chromosome
}
\details{
see \link{joint_seg_reads} for explanation. But effectively,
joint segmentation needs to happen per chromosome. So this function does the
segmentation, and the other is just a wrapper for all chromosomes.
}
