% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{library_from_cell}
\alias{library_from_cell}
\title{extract library ID from the typically formatted cell_ids}
\usage{
library_from_cell(cell_id)
}
\arguments{
\item{cell_id}{string of a cell_id or vector of cell IDs}
}
\value{
vector of the library ID(s) contained within
}
\description{
expecting cell IDs as AT21350-A143952A-R10-C37 with the second position being
the library ID.
}
