% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centromere_marking.R
\name{load_ucsc_centromeres}
\alias{load_ucsc_centromeres}
\title{load and prep default centromerefiles from UCSC}
\usage{
load_ucsc_centromeres(version = c("hg19", "hg38"), centro_file = NULL)
}
\arguments{
\item{centro_file}{NULL or string to the path if you download yourself}

\item{hg19}{boolean to target hg19 for loading}

\item{hg38}{boolean to target hg38 for loading}
}
\value{
tibble of parsed centromere spans
}
\description{
Cytoband files originally obtained from UCSC golden path genomes and parsed
to collect the total centromere spans.
}
\details{
hg19:
https://hgdownload.soe.ucsc.edu/goldenPath/hg19/database/cytoBand.txt.gz
hg38:
https://hgdownload.soe.ucsc.edu/goldenPath/hg38/database/cytoBand.txt.gz

TODO: can I stash the file if it's already loaded once?
}
