% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_differences.R
\name{make_cell_focused_matrix}
\alias{make_cell_focused_matrix}
\title{mostly internal for rearranging the pairwise DF to focus on a cell.}
\usage{
make_cell_focused_matrix(in_df, cell)
}
\arguments{
\item{in_df}{dataframe most likely from \code{\link[=pairwise_bin_difference]{pairwise_bin_difference()}}}

\item{cell}{string. Target cell ID to put as index cell}
}
\value{
tibble
}
\description{
This is useful because with non-redundant pair comparisons, the "last cell"
in the list is never the "index_cell", as it has already been compared ot all
others. But sometimes useful to still see it.
}
\details{
e.g.,
cells: A, B, C
comparisons would be: A-B, A-C, B-C
generate C focus with this function: C-A, C-B
or B focus: B-A, B-C
}
