% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_cellid_matrix}
\alias{make_cellid_matrix}
\title{pivot a cell_id dataframe to a wide matrix.}
\usage{
make_cellid_matrix(cell_df, name_col, val_col)
}
\arguments{
\item{cell_df}{some dataframe with a column of cell_id and values for each
of those cells.}

\item{name_col}{the column to be pivoted out}

\item{val_col}{the column to use as values in the matrix}
}
\value{
matrix
}
\description{
Different functions may produce a dataframe of counts or something per
cell_id for various attributes. This converts it to a matrix with rownames
of cell_id and columns of some feature of interest from the dataframe that
can be pivoted out.
}
\examples{
cell_df <- data.frame(
   cell_id = c("cell_1", "cell_1", "cell_1"),
   some_col = c("A", "B", "C"),
   some_val = 1:3
)

make_cellid_matrix(cell_df, name_col="some_col", val_col="some_val")

#       A B C
# cell_1 1 2 3

}
