% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_heatmap.R
\name{make_copynumber_heatmap}
\alias{make_copynumber_heatmap}
\title{create a ComlexHeatmap object of cell states}
\usage{
make_copynumber_heatmap(
  states,
  cell_id_order,
  clones,
  annos_df = NULL,
  labels_fontsize = 8
)
}
\arguments{
\item{states}{a table of cells states with a column of cell_ids, and other
columns of state bins (chrom_start_end).}

\item{cell_id_order}{a vector of cell_ids in the order that they should be
plotted in. Typically extracted from a ggplot of a tree (see
\code{\link[=cell_id_order_as_plotted]{cell_id_order_as_plotted()}}).}

\item{clones}{a table with at least cell_id and clone_id labels}

\item{annos_df}{optional table with a \code{sample_id} column and any number of
annotations to add for those sample IDs.}

\item{labels_fontsize}{int of label size for the chromosomes on the X and
clone IDs on the y of the heatmap.}
}
\value{
ComplexHeatmap::Heatmap object of states and any annotations.
}
\description{
This will create the typically DLP states heatmap of cells. It will add clone
labels and colors for each clade, and add any annotations given.
}
