% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_heatmap.R
\name{make_corrupt_tree_heatmap}
\alias{make_corrupt_tree_heatmap}
\title{create a complex heatmap object with a tree.}
\usage{
make_corrupt_tree_heatmap(phylo_tree, clones = NULL)
}
\arguments{
\item{phylo_tree}{a phylo object of a tree.}

\item{clones}{(optional) adds clones as OTUs, will eventually allow for color}
}
\value{
a ComlexHeatmap heatmap object with a tree.
}
\description{
Basically creates an empty heatmap with a tree added as a left side
annotation.
}
