% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_heatmap.R
\name{make_ordered_cell_id_annos}
\alias{make_ordered_cell_id_annos}
\title{expand an annotations dataframe to be for each cell.}
\usage{
make_ordered_cell_id_annos(annos_df, cell_id_order)
}
\arguments{
\item{annos_df}{a table of sample_id and columns of the desired annotations.}

\item{cell_id_order}{the order that the cell ids are plotted in, typically
extracted from a ggplot of a tree (see \code{\link[=cell_id_order_as_plotted]{cell_id_order_as_plotted()}})}
}
\value{
tibble of annotations for each cell_id.
}
\description{
Annotations are provided as per sample_id, this function expands those
annotations to each cell_id that they apply to. This is so ComplexHeatmap
can place the annotations for each cell in the heatmap.
}
