% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_believability.R
\name{map_states_to_letters}
\alias{map_states_to_letters}
\title{convert a vector of states to letters}
\usage{
map_states_to_letters(states)
}
\arguments{
\item{states}{vector of state values}
}
\value{
vector of letters corresponding to those states
}
\description{
The point is to then use measure string distance between cells. With raw
states double digit values would count as 2 characters and throw things off,
with standard R-functions to measure string differences.
}
