% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centromere_marking.R
\name{mark_bins_in_centromeres}
\alias{mark_bins_in_centromeres}
\title{mark edge of bin as within centromere or not}
\usage{
mark_bins_in_centromeres(
  reads_df,
  padding = 0,
  bin_location_column = "start",
  version = c("hg19", "hg38")
)
}
\arguments{
\item{reads_df}{tibble of read data}

\item{padding}{int of number of BP to add to each side of the centromere}

\item{version}{default 'hg19', or choose 'hg38' for locations of centromeres.}

\item{bin_locatation_column}{which column to use as the location of a bin.
popular choices are the start of a bin, middle of a bin, or and of a bin.}
}
\value{
input table, but with a boolean 'within_centro' column added (and
potentially other centromere information columns, if needed)
}
\description{
Can optionally specify a padding to mark locations of bins as being close
enough to a centromere. Often bins near centromeres are corrupt in their
state calls. In the past, we have filtered within 3 Mb of centromeres.
}
\details{
IMPORTANT! This adds the padding to each side of the centromere. So if you
specify a padding of 3 Mb, it will be within 3 Mb of the start and 3 Mb of
the end of the centromere.
}
