% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medicc_foreground.R
\name{medicc_profiles_to_foreground}
\alias{medicc_profiles_to_foreground}
\title{medicc profiles files to foreground assessment}
\usage{
medicc_profiles_to_foreground(
  medicc_profiles,
  phylogeny,
  cn_type = c("total", "allele")
)
}
\arguments{
\item{medicc_profiles}{imported medicc profiles. See
dlptools::read_medicc_profiles()}

\item{phylogeny}{phylo object of phylogeny generated by medicc. See
read_medicc_tree()}

\item{cn_type}{string, 'total' or 'allele' to indicate which type of medicc
output we are parsing. Characterization of changes is only done for total.}
}
\value{
table of tip data, their immediate parent states, and an inference
of whether a foreground event has happened  or not.
}
\description{
This function takes a medicc profile file (see
dlptools::read_medicc_profiles()) and a tree, and marks if changes in states
have occurred between parent nodes and tips.
}
