% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy.R
\name{mode_ploidy}
\alias{mode_ploidy}
\title{find the mode CN per chromosome, then mode across the chromosomes}
\usage{
mode_ploidy(
  bin_df,
  sample_col = "cell_id",
  cn_col = "state",
  chrom_col = "chr"
)
}
\arguments{
\item{bin_df}{dataframe of bin level data}

\item{sample_col}{string column name identifying samples}

\item{cn_col}{string column name for copy number states}

\item{chrom_col}{string column name for chromosomes}
}
\value{
tibble/dataframe of results by cell_id/sample
}
\description{
Done per chromosome first so that large chromosomes don't dominate the
result. It is critical that the input be even bin level data, or this makes
no sense to do. Measures of the copy number states need to be done in evenly
sized bins.
}
