% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converting_segments_and_reads.R
\name{old_segs_to_reads}
\alias{old_segs_to_reads}
\title{original function to split segments back to read bins}
\usage{
old_segs_to_reads(
  segs_df,
  bin_size = 5e+05,
  seg_start_col = "start",
  seg_end_col = "end"
)
}
\description{
kept for posterity. This splits into exact 500kb bins, which may extend
beyond end of chromosome, as original DLP did. New function is wildly faster
and more simple.
}
