% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_features.R
\name{ploidy_relative_cn_state}
\alias{ploidy_relative_cn_state}
\title{Could changes of state relative to ploidy.}
\usage{
ploidy_relative_cn_state(
  segs_df,
  sample_col = "cell_id",
  annotate_input = FALSE,
  return_matrix = FALSE
)
}
\arguments{
\item{segs_df}{a dataframe of CN segments}

\item{annotate_input}{boolean. return input dataframe annotating each}

\item{return_matrix}{boolean. Return a cell-by-feature matrix of counts.}
}
\value{
tibble/dataframe of counts
}
\description{
Marks CN segments as a gain or loss, relative to the mode ploidy of the
sample.
}
