% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_heatmap.R
\name{plot_combined_heatmap}
\alias{plot_combined_heatmap}
\title{generate a ComplexHeatmap::Heatmap image, either to console or file.}
\usage{
plot_combined_heatmap(
  total_hm,
  file_name = NULL,
  png_height = 1600,
  png_width = 2800,
  png_res = 144
)
}
\arguments{
\item{total_hm}{ComplexHeatmap::Heatmap of the combined tree and states.}

\item{file_name}{optional string of where to save a png image of the heatmap.}
}
\value{
ComplexHeatmap::draw or nothing if a file is written.
}
\description{
Will return an image of the generated heatmap or dump the heatmap to a file.
}
