% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_plots.R
\name{plot_gc}
\alias{plot_gc}
\title{visualize GC correction for a cell}
\usage{
plot_gc(reads_df, cellid, plot_choice = c("both", "raw", "corrected"))
}
\arguments{
\item{reads_df}{dataframe of reads like information.}

\item{cellid}{string of cell to plot}

\item{plot_choice}{string of plots to show. Default is both, alternatively "
'raw' for just the raw GC values or 'corrected' for just the corrected
values.}
}
\value{
ggplot2 object
}
\description{
Seeing how GC correction worked for a cell can give an indication if the
multiplier chosen by HMMCopy was appropriate to form integer copy numbers.
Also can just be informative to see what's going on behind the float copy
number to integer CN value.
}
