% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_differences.R
\name{plot_nnd_outlier_cells}
\alias{plot_nnd_outlier_cells}
\title{visualize cell nearest neighbour distance values}
\usage{
plot_nnd_outlier_cells(cell_diffs, outlier_cells)
}
\arguments{
\item{cell_diffs}{output dataframe of dlptools::pairwise_bin_difference()}

\item{outlier_Cells}{output dataframe of dlptools::find_outlier_cells()}
}
\value{
ggplot object
}
\description{
Generates a simple plot of each cell and highlights the min NN-distance for
each. Also highlights cells that are considered outliers.
}
