% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_heatmap.R
\name{pull_chr_from_col_name}
\alias{pull_chr_from_col_name}
\title{extract chromosome name from a bin name (chr_start_end)}
\usage{
pull_chr_from_col_name(col_name)
}
\arguments{
\item{col_name}{string of chr_start_end (column name in the internally used
state dataframe)}
}
\value{
string of the chromosome
}
\description{
an internal function for pulling information from a state matrix
}
