% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pull_info_from_cell_id}
\alias{pull_info_from_cell_id}
\title{generic extractor of info contained in cell ids}
\usage{
pull_info_from_cell_id(cell_id, info = c(NA, "sample", "library"))
}
\arguments{
\item{cell_id}{string or vector of cells id}

\item{info}{"sample" or "library" to pull}
}
\value{
vector of requested information
}
\description{
generic extractor of info contained in cell ids
}
