% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_heatmap.R
\name{pull_sample_id_from_cell_id}
\alias{pull_sample_id_from_cell_id}
\title{extract sample ID from the typically formatted cell_ids}
\usage{
pull_sample_id_from_cell_id(cell_id)
}
\arguments{
\item{cell_id}{string of a cell_id}
}
\value{
string of the sample ID contained within
}
\description{
expecting cell IDs as AT21350-A143952A-R10-C37 with the first position being
the sample ID.
}
