% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medicc_foreground.R
\name{read_medicc_profiles}
\alias{read_medicc_profiles}
\title{ingest a medicc profiles file}
\usage{
read_medicc_profiles(meddic_profiles_file)
}
\arguments{
\item{meddic_profiles_file}{file path}
}
\value{
tibble of file contents
}
\description{
This file is part of the standard medicc output files and contains all of
the input data and their states, plus the medicc generated state profiles
of the internal tree nodes.
}
\details{
Applies changes that are needed for other functions.
}
