% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_parsing.R
\name{read_medicc_tree}
\alias{read_medicc_tree}
\title{read medicc tree to phylo object.}
\usage{
read_medicc_tree(tree_file, drop_diploid = TRUE)
}
\arguments{
\item{tree_file}{file path to newick tree created by medicc}

\item{drop_dipliod}{boolean. Default TRUE. Remove diploid tip from tree.}
}
\value{
phylo object
}
\description{
Medicc forces the inclusion of a diploid ancestor for all cells and calls
that tip "diploid". This optionally drops that tip.
}
