% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converting_states_and_reads.R
\name{reads_to_segs_with_filters}
\alias{reads_to_segs_with_filters}
\title{convert reads table to segments table}
\usage{
reads_to_segs_with_filters(reads_df, with_mask = FALSE, mask_f = NULL)
}
\arguments{
\item{reads_df}{a table with standard reads data (e.g., created with
\link{import_dlp_files(file_type='reads')})}

\item{with_mask}{boolean on whether to filter mask==True read bins.}
}
\value{
tibble/dataframe with read bins organized into segment blocks.
}
\description{
Often the segments file given by the DLP pipeline is not what you want. What
you likely want is to do various filtering at the reads level, and then make
a segments file from that. This function will take a reads table and covert
it to a segments table.
}
\details{
Optionally can as it to remove reads that overlap with a mask region (see
\code{\link[=mark_mask_regions]{mark_mask_regions()}}), typically done with the meta_data/blacklist.txt file.
}
