% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_from_cell}
\alias{sample_from_cell}
\title{extract sample ID from the typically formatted cell_ids}
\usage{
sample_from_cell(cell_id)
}
\arguments{
\item{cell_id}{string of a cell_id or vector of cell IDs}
}
\value{
vector of the sample ID(s) contained within
}
\description{
expecting cell IDs as AT21350-A143952A-R10-C37 with the first position being
the sample ID.
}
