% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosome_layouts.R
\docType{data}
\name{SEG_CHROM_EVENT_LABELS}
\alias{SEG_CHROM_EVENT_LABELS}
\title{global of event labels for \code{\link[=mark_segs_chromosome_span]{mark_segs_chromosome_span()}}}
\format{
An object of class \code{character} of length 5.
}
\usage{
SEG_CHROM_EVENT_LABELS
}
\description{
Set as a global, as useful in other contexts
}
\keyword{datasets}
