% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_segment.R
\name{snap_segs_to_new_breakpoints}
\alias{snap_segs_to_new_breakpoints}
\title{alter existing segments to fit new breakpoints}
\usage{
snap_segs_to_new_breakpoints(jfit, input_mtx, input_cellids)
}
\arguments{
\item{jfit}{a \code{\link[jointseg:jointSeg]{jointseg::jointSeg()}} output object}

\item{input_mtx}{matrix. cell by bin state matrix}

\item{input_cellids}{vector. Cell ids of cells in the matrix.}
}
\value{
tibble. Cell segments that align with new breakpoints
}
\description{
alter existing segments to fit new breakpoints
}
