% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sort_df_by_cell_order}
\alias{sort_df_by_cell_order}
\title{sort a table given a vector of cell_ids}
\usage{
sort_df_by_cell_order(targ_df, cell_order)
}
\arguments{
\item{targ_df}{a table with cell_ids to sort}

\item{cell_order}{a vector of cell_ids in the desired order (e.g., pulled
from a ggplot of a tree)}
}
\value{
table that has been sorted
}
\description{
Typically used to sort a dataframe based on the plotted tip order to align
states heatmap/annotations/clone IDs to the plotted tree
}
