% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_heatmap.R
\name{sort_df_by_tip_order}
\alias{sort_df_by_tip_order}
\title{sort a table given a vector of cell_ids}
\usage{
sort_df_by_tip_order(targ_df, tip_order)
}
\arguments{
\item{targ_df}{a table with cell_ids to sort}

\item{tip_order}{a vector of cell_ids in the desired order (e.g., pulled from
a ggplot of a tree)}
}
\value{
table that has been sorted
}
\description{
Typically used to sort a dataframed based on the plotted tip order to align
states heatmap/annotations/clone IDs to the plotted tree
}
