% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{STATE_COLORS}
\alias{STATE_COLORS}
\title{alias of standard colors used in dlp plots}
\format{
An object of class \code{character} of length 13.
}
\usage{
STATE_COLORS
}
\description{
alias of standard colors used in dlp plots
}
\keyword{datasets}
