% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy.R
\name{weighted_ploidy}
\alias{weighted_ploidy}
\title{calculate sample ploidy with a weighted CN mean}
\usage{
weighted_ploidy(
  cn_df,
  sample_col = "cell_id",
  start_col = "start",
  end_col = "end",
  cn_col = "state",
  chrom_col = "chr"
)
}
\arguments{
\item{cn_df}{dataframe input of segmented (or not) sample copy numbers}

\item{sample_col}{string column name identifying samples}

\item{start_col}{string column name identifying start of segments}

\item{end_col}{string column name identifying end of segments}

\item{cn_col}{string column name for copy number states}

\item{chrom_col}{string column name for chromosomes}
}
\value{
tibble of per-sample ploidy estimates
}
\description{
CN is first weighted by the length of its segment, a weighted mean of these
segments for each chromosome is made, then the mean across all chrosomomes
is made for each sample.
}
\details{
In this way, the CN contribution to the ploidy is based on it's length, and
each chromosome contributes equally to the overall sample ploidy.
}
