% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_features.R
\name{wu_features_to_matrix}
\alias{wu_features_to_matrix}
\title{convert feature counts based on Wu et al. to a matrix of crossed features.}
\usage{
wu_features_to_matrix(feat_count)
}
\arguments{
\item{feat_count}{default count dataframe output of \code{\link[=extract_wu_features]{extract_wu_features()}}.}
}
\value{
matrix object of counts. Columns are crossed feature categories,
rows are cell ids.
}
\description{
Takes the count matrix from \code{\link[=extract_wu_features]{extract_wu_features()}} and returns a matrix
object. \code{\link[=extract_wu_features]{extract_wu_features()}} offers to return a matrix, but sometimes
you don't want to redo the extraction.
}
