% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosome_layouts.R
\name{add_centromere_locations}
\alias{add_centromere_locations}
\title{add centromere information to reads by chromosome}
\usage{
add_centromere_locations(
  reads_df,
  centro_file = NULL,
  version = c("hg19", "hg38")
)
}
\arguments{
\item{centro_file}{NULL or string to the path if you download yourself}

\item{hg19}{boolean to target hg19 for loading}

\item{hg38}{boolean to target hg38 for loading}
}
\value{
input table with centromere information added by chromosome
}
\description{
See dlptools::read_and_prep_ucsg_cenrtomeres() for details of file origins.
}
