% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wu_cn_features.R
\name{add_wu_seg_state_bins}
\alias{add_wu_seg_state_bins}
\title{see dlptools::extract_wu_features}
\usage{
add_wu_seg_state_bins(
  segs_df,
  state_bin_max = 5,
  bin_breaks = c(5e+06, 1e+07 + 1)
)
}
\value{
tibble
}
\description{
see dlptools::extract_wu_features
}
