% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggo_tree.R
\name{build_aggo_tree}
\alias{build_aggo_tree}
\title{build tree with AGNES}
\usage{
build_aggo_tree(
  reads_df,
  cut_k = 8,
  state_col = "state",
  sample_col = "cell_id",
  by_ploidy_change = FALSE
)
}
\arguments{
\item{reads_df}{data.frame. bin based reads data}

\item{cut_k}{int. level to cut the tree at.}

\item{state_col}{str. Column to use for the clustering. Default: "state"}

\item{sample_col}{str. Column of the sample labels. Default: "cell_id"}

\item{by_ploidy_change.}{bool. Use CN change from cell mode ploidy as the
feature to cluster sample by. This can help group clones that are WGD of
lower ploidy clones.}
}
\value{
list. Two elements. $phylo: the tree; $clones: tibble of clone ids
of tip labels based on tree cutting.
}
\description{
Agglomerative hierarchical clustering is a method of clustering that can
produce a tree-like structure. Here, we construct a distance matrix based,
typically, on copy number states of bins, then feed that distance matrix
to AGNES clustering.
}
\details{
This function will also do a preliminary tree cutting to provide groups
within the tree, which could be interpreted as copy-number clones of cells.
You can always re-cut the returned tree with \code{\link[stats:cutree]{stats::cutree()}}, which is all
that is used internally here.
}
