% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plot.R
\name{build_left_annot}
\alias{build_left_annot}
\title{builds the left-side annotations of the cells}
\usage{
build_left_annot(
  anno_df = NULL,
  anno_cols_list = list(),
  clones_df = NULL,
  only_largest_clone_group = FALSE,
  labels_fontsize = 8,
  clone_palette = NULL
)
}
\description{
builds the left-side annotations of the cells
}
