% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_plots.R
\name{chip_plot}
\alias{chip_plot}
\title{Plot metrics across the chip}
\usage{
chip_plot(metrics_df, targ_val)
}
\arguments{
\item{metrics_df}{dataframe of metrics for a run.}

\item{targ_val}{the value of interest for plotting. E.g.,
experimental_condition, quality, total_reads.}
}
\description{
Map of a metric as observed in the layout of the DLP chip.
}
