% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_features.R
\name{extract_ploidy_cn_feature}
\alias{extract_ploidy_cn_feature}
\title{Could changes of state relative to ploidy.}
\usage{
extract_ploidy_cn_feature(
  segs_df,
  sample_col = "cell_id",
  annotate_input = FALSE,
  return_matrix = FALSE
)
}
\arguments{
\item{segs_df}{dataframe. CN segments}

\item{sample_col}{string. Name of the column with cell_id/other sample name}

\item{annotate_input}{boolean. return input dataframe annotating each}

\item{return_matrix}{boolean. Return a cell-by-feature matrix of counts.}
}
\value{
tibble/dataframe of counts
}
\description{
Marks CN segments as a gain or loss, relative to the mode ploidy of the
sample.
}
