% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_features.R
\name{extract_segment_position_feature}
\alias{extract_segment_position_feature}
\title{count the segment-span-on-chromosome event types.}
\usage{
extract_segment_position_feature(
  segs_df,
  sample_col = "cell_id",
  annotate_input = FALSE,
  return_matrix = FALSE,
  ...
)
}
\arguments{
\item{segs_df}{dataframe of CN segments}

\item{sample_col}{string. Name of the column with cell_id/other sample name}

\item{annotate_input}{boolean. return input dataframe annotating each
segment.}

\item{return_matrix}{boolean. Return a cell-by-feature matrix of counts.}
}
\value{
tibble/dataframe of counts
}
\description{
Critical to this fuction is \code{\link[=mark_segs_chromosome_span]{mark_segs_chromosome_span()}}. It is
important to read and understand that function and its arguments.
}
\details{
This function basically just calls \code{\link[=mark_segs_chromosome_span]{mark_segs_chromosome_span()}}
and summarizes the results. Arguments can be passed to that underlying
function. Passing no arguments means you are happy with the defaults. See
\code{\link[=mark_segs_chromosome_span]{mark_segs_chromosome_span()}} to understand what the defaults are.
}
