% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_features.R
\name{extract_segment_sizes}
\alias{extract_segment_sizes}
\title{sizes of the segments}
\usage{
extract_segment_sizes(segs_df, sample_col = "cell_id")
}
\arguments{
\item{segs_df}{dataframe. copy number segments for samples.}

\item{sample_col}{string. Name of column with cell/sample names}
}
\value{
dataframe. sample ids and all observed segment sizes.
}
\description{
This function is as simple as it sounds, end - start.
}
\details{
Used as a setup for extracting process based features \emph{á la}:
\itemize{
\item Macintyre et al. 2018
\item Drews et al. 2018
}

Really, something like this should be used to generate values that then you
define categories for to count occurrences.

Related functions include: \link{extract_extract_changepoint} and
\link{extract_breakpoints}
}
