% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plot.R
\name{fetch_heatmap_color_palette}
\alias{fetch_heatmap_color_palette}
\title{grab colors for various hm possibilities.}
\usage{
fetch_heatmap_color_palette(state_col, states_df)
}
\description{
Standard colors used by Signals and other people from the DLP world.
}
