% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_heatmap.R
\name{format_states_for_hm}
\alias{format_states_for_hm}
\title{prepare a states table for heatmap plotting}
\usage{
format_states_for_hm(states, tree_tips)
}
\arguments{
\item{states}{table of bin rows, cell ID columns with state calls}

\item{tree_tips}{vector of the order that cell ids should be in (ideally
pulled from a ggplot tree object)}
}
\value{
matrix of state calls.
}
\description{
Will take the states table (bin rows, cell ID columns) and sort it by the
plotted tree order, convert to a matrix (which complexheatmap needs), and
sort it by chromosome. It also converts state 11 to 11+, which is what it
really is.
}
