% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plot.R
\name{generate_state_hm}
\alias{generate_state_hm}
\title{creates a complex heatmap of a given matrix of states.}
\usage{
generate_state_hm(
  states_mat,
  phylo,
  labels_fontsize = 8,
  plot_cols = STATE_COLORS,
  left_annot = NULL,
  legend_11plus = FALSE
)
}
\description{
creates a complex heatmap of a given matrix of states.
}
