% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_chr_used_in_chroms}
\alias{is_chr_used_in_chroms}
\title{detect if chromosomes labels include "chr"}
\usage{
is_chr_used_in_chroms(chrom_vec)
}
\arguments{
\item{chrom_vec}{vector of chromosome labels, probably pulled from a
dataframe of cn information.}
}
\value{
boolean. TRUE if "chr" found in the strings.
}
\description{
detect if chromosomes labels include "chr"
}
