% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosome_layouts.R
\name{load_chrom_info_file}
\alias{load_chrom_info_file}
\title{loading UCSC chromosome length files}
\usage{
load_chrom_info_file(version = c("hg19", "hg38"))
}
\arguments{
\item{version}{default "hg19", can also load "hg38"}
}
\value{
tibgble of chromosome, total length, etc.
}
\description{
loading UCSC chromosome length files
}
