% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing.R
\name{load_mask_file}
\alias{load_mask_file}
\title{internal to control mask file loading.}
\usage{
load_mask_file(mask_f = NULL)
}
\arguments{
\item{mask_f}{path to the mask file to load.}
}
\description{
internal to control mask file loading.
}
