% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_hm.R
\name{make_clone_palette}
\alias{make_clone_palette}
\title{generate a color palette for clone labels}
\usage{
make_clone_palette(clone_ids)
}
\arguments{
\item{clone_ids}{unique vector of clone labels}
}
\value{
named vector of hex colors, names are clone labels
}
\description{
generate a color palette for clone labels
}
