% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_differences.R
\name{pairwise_bin_difference}
\alias{pairwise_bin_difference}
\title{metric of pairwise differences between two cells.}
\usage{
pairwise_bin_difference(bin_df, cells = c(), min_seg_length = 2500000)
}
\arguments{
\item{bin_df}{a dataframe of read bins with states. Expected columns of:
cell_id, chr, start, end, state}

\item{cells}{optional vector specifying cells to compare. If it's blank, all
cells are compared. If it's 1 cell, then that one cell is compared to all
others. If it's 2 or more, then just the specified cells are compared to each
other.}

\item{min_seg_length}{double. This is the minium length of matching segment
bins to use when measuring similarity.}
}
\value{
tibble of cell pairs and metrics about their differences.
}
\description{
Inspired by MSKCC SPECTRUM paper. Bins are aligned between two cells and
marked for if they have the same state Then segments of matching and
non-matching runs of bins are found (filtering those smaller than a
specified minimum). These segments are then re-split into 500kb bins and the
difference becomes the number of matching bins divided by the number of
considered bins.
}
\details{
This function is slow, and the number of pairwise comparisons grows quickly.
Dramatic speed improvements can be had by setting up a parallel plan for
furrr like so:

future::plan(future::multicore, workers=N_CORES_YOU_WANT)

Example for 100 cells, which is 4950 pairs, this function will take 2 minutes
with 4 cores.
}
