% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreground_plots.R
\name{plot_bg_state_highlight}
\alias{plot_bg_state_highlight}
\title{generate a plot of just the background states, making the foreground white.}
\usage{
plot_bg_state_highlight(states_df, file_name)
}
\arguments{
\item{states_df}{bin based dataframe that has foreground events marked.}

\item{file_name}{name for the resulting png.}
}
\description{
These are the actual state values of the bins in background bins.
}
