% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_differences.R
\name{plot_nnd_outlier_cells}
\alias{plot_nnd_outlier_cells}
\title{visualize cell nearest neighbour distance values}
\usage{
plot_nnd_outlier_cells(pairwise_diffs, nn_cells, outlier_cells)
}
\arguments{
\item{pairwise_diffs}{dataframe of \code{\link[=pairwise_bin_difference]{pairwise_bin_difference()}}}

\item{nn_cells}{dataframe of \code{\link[=find_nearest_neighbours]{find_nearest_neighbours()}}}

\item{outlier_Cells}{dataframe of \code{\link[=find_outlier_cells]{find_outlier_cells()}}}
}
\value{
ggplot object
}
\description{
Generates a simple plot of the nearest neighbour distance for each cell,
which is highlighted with the point. Outlier cells are indicated by colour.
A violin plot is used to show distance to all other cells.
}
