% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_plots.R
\name{plot_read_bins_basic}
\alias{plot_read_bins_basic}
\title{create a tile plot of read state calls.}
\usage{
plot_read_bins_basic(reads_df)
}
\arguments{
\item{reads_df}{a table of reads data}
}
\value{
ggplot object
}
\description{
builds a basic ggplot with geom_tile on a reads df.
Only expects columns of cell_id, start, state, chr
}
