% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plot.R
\name{plot_state_hm}
\alias{plot_state_hm}
\title{main hm building function}
\usage{
plot_state_hm(
  states_df,
  state_col,
  phylogeny = NULL,
  anno_df = NULL,
  anno_colors_list = list(),
  clones_df = NULL,
  anno_columns = NULL,
  clone_column = NULL,
  color_tree_clones = FALSE,
  only_largest_clone_group = FALSE,
  file_name = NULL,
  labels_fontsize = 8,
  ...
)
}
\description{
anno_cols_list: list(Passage=c(\code{3}: #123456))
}
