% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pull_info_from_cell_id}
\alias{pull_info_from_cell_id}
\title{generic extractor of info contained in cell ids}
\usage{
pull_info_from_cell_id(cell_id, library_id = FALSE, sample_id = FALSE)
}
\arguments{
\item{cell_id}{string or vector of cells id}

\item{library_id}{boolean to extract library IDs}

\item{sample_id}{boolean to extract sample IDs}
}
\value{
vector of requested information
}
\description{
generic extractor of info contained in cell ids
}
