% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converting_states_and_reads.R
\name{reads_to_segs}
\alias{reads_to_segs}
\title{convert reads table to segments table}
\usage{
reads_to_segs(reads_df)
}
\arguments{
\item{reads_df}{a table with standard reads data (e.g., created with
\link{import_dlp_files(file_type='reads')})}
}
\value{
tibble/dataframe with read bins organized into segment blocks.
}
\description{
Often the segments file given by the DLP pipeline is not what you want. What
you likely want is to do various filtering at the reads level, and then make
a segments file from that. This function will take a reads table and covert
it to a segments table.
}
