% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converting_states_and_reads.R
\name{rle_states}
\alias{rle_states}
\title{convert states to run length encoding}
\usage{
rle_states(states)
}
\arguments{
\item{states}{really and vector of values.}
}
\value{
vector of integers
}
\description{
takes a vector of numbers (e.g., states) and returns a numeric group
value indicating the a group for each run of the same value.
c(5,5,5,6,6,5,5,5,2) -> 1 1 1 2 2 3 3 3 4
}
