% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converting_segments_and_reads.R
\name{segs_to_reads}
\alias{segs_to_reads}
\title{split segments into bins of a requested size.}
\usage{
segs_to_reads(
  segs_df,
  bin_size = 5e+05,
  seg_start_col = "start",
  seg_end_col = "end"
)
}
\arguments{
\item{segs_df}{dataframe (or similar) of segment dat}

\item{bin_size}{width of bins to split into. Default is standard 500kb}

\item{seg_start_col}{name of the column that indicates the start of a segment}

\item{seg_end_col}{name of the column that indicates the end of a segment}
}
\value{
input frame split into bins
}
\description{
Takes a dataframe with segment start and end columns and returns a dataframe
with those segments split into individual bins. All segment information
applied to the generated bins.
}
\details{
Warnings issues if requested bin size is bigger than some segments or if
segments can't be split evenly into the bins.

Bin end will not exceed a segment end. Depending on what is input and
requested, this can lead to one smaller bin at the end of the segment.
}
